UITextView is getting closer to functional. It overlays a fresh NSTextView on to the whole area when a touch is detected within its bounds, swaps in the right stuff, shares the same text storage objects, etc, so there's only ever the one NSTextView on screen as long as the responder stuff is setup right and only one UIResponder is ever active at a time. (Of course, bugs could perhaps circumvent this... :P) This is a bit different than a field editor which keeps a object around and reuses it, but I don't think it's so terribly expensive to just create it as needed.
It even supports secure text entry mode via some seriously fancy glyph substitution stuff, IMO. :) Sure, I could have used the NSSecureTextField, except that's only a field and technically even UITextView needs to support secure entry mode. (Although honestly I've not yet tried it in iOS so I have no idea if it even allows it there on a text view. The UITextInputTraits protocol requires it, though, which UITextView and UITextField are both said to conform to so I totally went all the way. Plus, it was interesting to figure out.)

I should be able to reuse UIText for UITextField's re-implementation so that it's all consistent across the board. I will have to handle things like enter/return/tab differently for a field, though. Also, no delegates are being called yet, but it's getting there.

One major downside to this whole approach is of course that the NSTextView is going to sit on top of the other views regardless of their actual ordering in the hierarchy. Plenty of pain went into trying to make that not be the case, but it is either not possible or I just don't know enough of the fancy tricks and hacks AppKit does with layers.

Another issue is that normally you can do things like drag-drop text into an NSTextView, but because the view doesn't exist until you make the UITextView the first responder, it doesn't work. I suspect that with some more hacking on the event delivery stuff I could come up with a reasonable way to pull that off but I imagine it's not all that important in the long run.

A final issue (while I'm thinking about this) is I really don't know how any of this will interact with accessibility - but we'll cross that bridge later. :)

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@3769 d3437ac4-8b1b-0410-be28-b75ab53cb89f