Rearchitected the text views/NSView stuff a bit. It now keeps the NSViews around for the entire life cycle of the UITextLayer, but it adds/removes it from the screen semi-intelligently. The frames are now kept in sync using UIView notifications rather than KVO as KVO was proving to be problematic to keep straight.
The really cool thing, though, is that the NSView's now have a custom hitTest: method on them which results in their being properly occluded when the UIView that is hosting them is behind another UIView!  It's late and I'm tired, but that may be the coolest thing ever. As far as the user is concerned, I think these NSViews now operate entirely as expected both from the point of view of UIKit as well as AppKit. It's pretty freaking epic.

I am awesome.

And tired.

But mostly awesome.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@3782 d3437ac4-8b1b-0410-be28-b75ab53cb89f