I need to come up with some other way to dismiss the popovers in response to external clicks.
On the real iOS, I suspect that what is basically happening is that whenever there's a tap anywhere outside of a popover, something is checking to see if it should cause a dismiss or not (this would be how the pass through view stuff would work and the specialness of popovers presented from a button bar item). Initially, our popover was making an assumption that any clicks outside of it should dismiss it - which isn't technically true. (I have a suspicion that on iOS, a popover creates a full screen window that it then overlays on top of everything else in order to capture clicks and handle the thing that allows you to present modal views inside a popover. This might be tricky to simulate, but I'll need to think about it a bit. The real iOS doesn't have a situation where a popover can actually extend off off the side of the screen like we want to allow... so... yeah..)

The trouble with Ostrich was that when it went to present the result of "Show Conversation", that resulted in a new popover being created. That made the previous one which housed the UIActionSheet to lose it's status as the main window and that caused it to try to dismiss itself immediately. Previous to the epic refactor, those menu items weren't acted on until the menu disappeared, I believe, which is why we didn't notice this problem before. The new system acts on the menu items before it actually dismisses the menu - that means that for a brief period of time, there's actually two popovers present (but not necessarily visible).

Any menu item which resulted in bringing up a new popover would probably have triggered this bug.

Since I've now disabled this, you have to actually close popovers if you want them to go away in the usual "clicked outside the window" way. This is obviously not the correct solution, but it works as a workaround until I get this situation fixed up somehow.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@3962 d3437ac4-8b1b-0410-be28-b75ab53cb89f