Whew. Selection should now be correct in every case I can think of. This change is hard to describe, but what it does is ensures the selection rectangles extend all the way to the edge of the line when the selection wraps across a partial line. So:
This is a
multiline example.

If the selection started on "is" and extended to the end of "multiline", the first line's selection rect needed to be clamped on the far right side of the view - not just the end of the word "a". This was previously not the case. It now is. This was all made much more difficult by the fact that right-to-left languages need the snapping to occur on the *left* side instead and, of course, in a string with both left-to-right and right-to-left languages in it, it still needed to act correctly when crossing those boundaries. I think I got it to work, though, so... yay! It feels quite natural now, IMO.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@3996 d3437ac4-8b1b-0410-be28-b75ab53cb89f