Implemented touch canceling when a view is removed from its superview.
After some pretty heavy sleuthing, I think I found one of the mysterious-but-rare Ostrich crashers which happen if you're clicking on something in a table cell and the table reloads under you at the same moment. Very difficult to reproduce, so I can't be sure yet if this is it.

The reload causes cells to get destroyed and recreated (or sometimes not recreated at all) and if a touch was active from within that destroyed cell at the time, fun crashy times could happen.

The problem was that clicking on the content view of the cell, for instance, would keep the content view around because UITouch would have retained it as the starting view for the touch, but UITouch doesn't retain the entire view hierarchy - so the content view's superview (the cell itself) was not retained and should be getting removed and ultimately deallocated. At some point later, when you let the mouse go, it'd try to deliver touch events to the content view (touchesEnded) which ends up passing that on to it's superview (the cell) which no longer exists and sometimes caused a crash.

There's probably more to this because UIView is supposed to be setting all of it's subview's superview to nil when it's destroyed (this is done in -dealloc) and I haven't quite figured out how come that nil-ing process didn't prevent events from exploding in this case because it should never even have been able to deliver the touch ended events to the superview of the content view if the superview was nil! But that'd have just masked the problem because the real UIKit appears to cancel (or abort) touches when a view is removed that has any touches active in it. If any nil magic had been working, it still wouldn't have functioned the same way. Best to just abort the touch entirely the way UIKit does it as that is probably the most reliable behavior in this situation.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4280 d3437ac4-8b1b-0410-be28-b75ab53cb89f