This is a very subtle adjustment to how touch canceling events work.
This was difficult to figure out, but these changes should now allow for certain touch behaviors to feel more... Mac-y under certain circumstances. It also eliminates the ugly deferred mouse exit events being sent to the view when a touch was cancelled which have always felt potentially dangerous.

One thing this change enables is for more Mac-like drop down menus. The new action button on the Ostrich profile view (not yet checked in as of this writing), can now work like this:  Click, action menu pops up, drag to desired action, release mouse to perform the action. While doing that operation, the action menu still displays in its "pressed" state. After releasing the mouse, the action menu gets a cancelled touch event which resets it back to the non-pressed state at the same time that the desired action is performed.

Prior to this change, this sort of thing was not possible as the cancelled touch triggered by displaying the popup menu would immediately reset the button's state back to the non-pressed state, even though you might still be holding down the mouse. So the menu could still appear in that case, but you'd never actually see the button's pressed state. This felt very unnatural to me. Now we can support the click-drag-select usage pattern that is very natural on normal OSX drop down menu controls while retaining the button's "pressed" state while that process is in progress.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4282 d3437ac4-8b1b-0410-be28-b75ab53cb89f