The UIScrollView dragging delegates should now be called in all the ways that a scroll view can be scrolled by the user.
It works by setting up a timer whenever a scroller changes the view's contentOffset. This happens in a variety of ways - when the user clicks in the scroller's gutter, when they drag the knob, or when they use the mouse wheel. Since all these methods had varying meanings of when a "drag" begins and ends, and a user could rapidly switch between two modes (like, dragging the knob, releasing, and then immediately scrolling a bit with the wheel), it seemed better to come up with a single way to handle it that's consistent, simple, and easy to verify.

Any scroll attempt from the user will put the view into a "dragging" mode which starts a timer. Further scrolling within a small time interval will keep the scroll view in the dragging mode. Once scrolling movement ceases, the timer eventually expires, and it leaves the dragging mode and reports that to the delegate. If the user is still holding down the mouse button after dragging the knob, but simply not moving it, that is irrelevant - the dragging state is based on timing. Any further movement will simply restart the dragging mode and it proceeds as if the user had initiated a new scroll request. In practice I don't think this will make a significant difference to client code.

The primary differences between this implementation and the real UIKit is that the end drag message is delayed longer than it would be on iOS, where it is delivered the moment the user lifts their finger. We also never report any deceleration since momentum scrolling on OSX automatically sends "fake" wheel scroll events, they are not documented as being distinguished from a regular scroll event, so I can't easily mimic the deceleration stuff.

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4310 d3437ac4-8b1b-0410-be28-b75ab53cb89f