zoomScale is now a calculated property and does not use an ivar. It appeared to me that the real UIKit took the "a" field from the transform and used that as the zoomScale regardless of if it was tampered with or not. So, in the interest of consistency, I shall do this as well until further notice.
The docs suggest that zooming doesn't work unless both of these delegates are implemented:
-viewForZoomingInScrollView:
-scrollViewDidEndZooming:withView:atScale:

However empirical evidence says that may not be the case as scrollViewDidEndZooming:withView:atScale: doesn't seem to exist anywhere in the Ostrich code base and yet the picture viewer has historically worked okay on iOS. Of course, at the same time, when I tried to use our old IFCenteredScrollView directly in a simple stand-alone test case, it crashed horribly on iOS 4.2. Something must have gotten changed somewhere; maybe it changed in the refactor or when that code was first brought over to Ostrich. I don't know...

git-svn-id: svn://dev.iconfactory.net/Projects/Software/UIKit/trunk@4542 d3437ac4-8b1b-0410-be28-b75ab53cb89f