Simplified -convertRect:toView: / fromView:
These operations happen alot. The existing implementation was converting *two* points, the top-left and lower-right corners of the rect.  Since the size of the rect should not be affected by these operations, the relationship between the two points will not change and it stands to reason that only the origin needs to be converted.