BugFix: Better implementation of common init.
NSCoding presents us with a bit of a dilemma.  In order to implement it properly, we need to have two, parallel chains of constructors -- The normal one and -initWithCoder:, which is the back-door...  each of these constructors is expected to call it's respective super.  So, this means that there is a lot of common code that can be duplicated between the two sets of constructors, which violates DRY.  The normal solution would be to push the code out into a method, shared between the two, but this runs into difficulty with subclassing.  The simple solution is to give each "common init" method a unique signature by including the name of the class in the method name.