Fixed a problem where touches got cancelled if the original view that was the source of the touch is removed before the touch ends.
This did not work well with gesture recognizers because the touch would
get cancelled and thus the entire gesture would go with it. On a UITable
View that did not reuse cells, this caused serious problems scrolling. I
t was not noticed before because previously scroll events were treated a
s new discrete touches each time they occurred, so as the table scrolled
 new cells would be under the mouse and thus it masked the problem.