Now creating contentView, etc. on demand.
Not sure if the real thing does this or not, but it may. A subclass of
UITableViewCell that doesn't put stuff on the contentView directly but
instead just puts its own view on top of the cell seems to run into
problems with the contentView stealing taps/clicks unless it's made
on-demand and sorted properly. This may not be a good idea in the long
run but it seems to fix something for me at the moment.