Implemented backgrounding support and UIApplicationState switches.
Chameleon now switches the applicationState property on UIApplication
appropriately. In addition to the previous behavior of going to inactive
when the app left the foreground, it now also goes inactive when the
screen goes to sleep (if it happened to be foregrounded). The big one,
though, is that Chameleon now supports "backgrounding" which is
activated when the machine is about to go to sleep. When this happens,
Chameleon begins a switch to UIApplicationStateBackground and also
attempts to stall the machine's sleep until background tasks have
finished. When the machine wakes up, it will resume in
UIApplicationStateInactive (as per UIKit docs).

Conflicts:

	UIKit/Classes/UIApplication.h
	UIKit/Classes/UIApplication.m