Improved user experience when using task completion.
It will now block for a brief time to allow the tasks to finish as it is
quitting. If it is unable to complete the tasks in that time, it'll show
the alert. This is to prevent a flashing alert right when the user quits
the app that's too brief to read (since most task completions are very
quick server syncs, etc.). Additionally, if the alert does appear, it
will remain on screen for about 2 seconds even if the tasks complete
before then. This allows the user time to read the alert and not be so
startled by it. Closing the alert works as before and will quit the app
immediately and cancel any unfinished tasks.