Fix strange behavior when context stack is empty.
The following patch accomplishes three things:

1) It pushes the current graphics context (from [NSGraphicsContext
currentContext]) whenever the internal context stack is empty. This
allows us to return to the original system context ever before we
started to populate our internal stack.

2) When popping the stack it checks to see if we've returned to the
original context (before the stack was populated) and empties the
stack ready to begin the cycle all over again.

3) When querying UIGraphicsGetCurrentContext and the internal context
stack is empty, we return the current system context
([NSGraphicsContext currentContext]).

Note: I didn't leave the system context permanently on the internal
stack (see 2 above) because I'm not 100% sure if the system will
change the context when rendering to a different window or an
internal buffer that may not persist through the life of the
application.