Completely redid UIImage to support multi-resolution images.
Unlike the real iOS (as of iOS 5 to my knowledge), my UIImage
internally stores multiple representations of images if they are found
so that retina support can work. On iOS, you either have a retina
device or not so it can make that choice up front, but on OSX you can
draw a window to another screen, change your screens prefs, etc. and
the retina status of a window or image can change on the fly. This
infrastructure allows it to choose the correct image for the situation
it's in at the time all automatically. You don't have to do anything.
It also uses CGImageSource to sort of lazy-load image information so if
your app has retina assets but never runs on a retina display, it won't
load any more of the retina image files than it needs to determine
their sizes and existence. Also moved the NSImage support out of the
main UIImage code and into a category as well as ensured it supports at
least 1x and 2x images when transferring between UIImage and NSImage.