Much like UIImage, UIColor needed multiple internal representations in order to work correctly when switching between retina and non-retina displays.
Obviously normally color would not need this, but the pattern images
proved difficult to get right and this seemed to be the best way. If a
color is created using just a normal solid color, only one
representation is used so there's not a significant increase in
complexity or memory use in the normal case.