Events, ARC, and more
Sorry for everything lumped into one but it’s been a long time and I
didn’t do a good job keeping our internal branch (which is stuck in
svn) in sync with the this one. I know I know I’m doing it wrong! :P
This includes a bunch of modernization of the codebase including a
switch to ARC (which many of you have already done). It also includes
some rather significant changes to how events get from AppKit to UIKit,
gesture recognizers, responder chain changes (attempts to integrate
between NSResponder and UIResponder), and probably a bunch of other
stuff I’ve forgotten about. Due to a heavy focus on iOS work lately,
this code has not been extensively used or tested but it was well past
time to get it out there.